refactor: api 명세에 맞게 타입 수정

+++ android/app/src/main/java/com/app/edonymyeon/data/dto/response/CommentsResponse.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class CommentsResponse(
     @SerialName("commentCount")
-    val commentCount: Long,
+    val commentCount: Int,
     @SerialName("comments")
     val comments: List<CommentDataModel>,
 )
