+++ android/app/src/main/java/com/app/edonymyeon/presentation/uimodel/ReactionCountUiModel.kt
@@ -6,5 +6,5 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class ReactionCountUiModel(
     val viewCount: Int,
-    val commentCount: Long,
+    val commentCount: Int,
 ) : Parcelable
